[US_NY] Infer entries to ensure that people are always in the right location on file dates (Recidiviz/recidiviz-data#50940)
## Description of the change

There are some rare cases where a person will be listed in an
UNDER_CUSTODY file with a facility value that can't be explained by the
data we're pulling in to build IPs. This change infers entries to
facilities in these cases to ensure that people are always in whatever
location the UNDER_CUSTODY file says on the file date, since we're
supposed to take the raw snapshot information in the table as the source
of truth, not the derived movements and periods. If someone is in the
2025-01-01 UNDER_CUSTODY file with facility = A but no information about
how they got there (so no admissions or movements into facility A), we
need to infer a movement into facility A on 2025-01-01 so that the
population counts are correct.

This comes up most commonly when people have non-monotonic transfer
dates over multiple UNDER_CUSTODY files. Sometimes, someone will be
transferred to a facility with the correct transfer date (facilda)
listed, but then in the next UNDER_CUSTODY file they'll be back in their
original facility, with a facilda of whatever date they originally
entered that facility. For example:
- 2025-01-01 UNDER_CUSTODY file: shows person A in facility 1 with
facilda 2024-01-01. This means that the person entered facility A on
2024-01-01 and is still there on 2025-01-01.
- 2025-02-01 UNDER_CUSTODY file: shows person A in facility 2 with
facilda 2025-01-15. This means that the person was transferred to
facility 2 on 2025-01-15 and is still there as of 2025-02-01.
- 2025-03-01 UNDER_CUSTODY file: shows person A in facility 1 with
facilda 2024-01-01. This means that the person entered facility A on
2024-01-01, which we already know from the 2025-01-01 file. However, it
also means that the person was in facility 1 on 2025-03-01, despite
there not being any indication of why/when they re-entered from facility
2. Therefore, we need to infer an entry to facility 1 on 2025-03-01,
because otherwise, the IP in facility 2 would never end even though we
know for a fact that the person had returned to facility 1 as of
2025-03-01.

This is still a relatively rare case, and seeems to be the result of
data entry errors where facilda dates don't get updated after temporary
transfers.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 36776e720e25118a9bc96f280317c9286d915e8e