[tools] Match chunked file format when deprecating raw files (Recidiviz/recidiviz-data#50890)
## Description of the change

during the utah reimport i noticed i had reimported files that had been
deprecated in the operations db. it turns out our script doesnt support
chunked file format so anytime we've been deprecating a chunked file we
haven't been moving it from storage. we use different logic when
deprecating files in the ingest bucket so that didnt need to be updated.

i thought about adding both formats for all file tags in case there were
to be a case where a file was chunked and then became non-chunked, but
this script is already kind of slow and i didn't want to slow it down on
every run for an edge case that _probably_ won't happen.

there are only two other chunked file tags so i will check their entries
in the operations db and also check prod to see if any other files need
to be moved

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#48615

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0abe400705dab9c79b4e07eca1ae14d70fead3d6