drop Firestore ETL batch to 50 (Recidiviz/recidiviz-data#51070)
This got our still-problematic residents file for TN to succeed for me
in a local run. If we keep dividing by 2 then mathematically speaking we
can continue doing this forever!

(We are working on a more durable solution but I am hoping this will be
a temporary stopgap to keep data fresh in the meantime)

GitOrigin-RevId: 010f31c064debbf0a2386d0d77e0ef22b8c42e64