[Sentences v2] Add test that finds all v1 references that needed to be migrated, broken down by product (Recidiviz/recidiviz-data#50939)
## Description of the change

Adds a series of tests in `sentence_v1_migration_burndown_test.py` that
explicitly outlines the "bad" edges that a metric export view depends
on, where a bad edge is an edge between a deprecated v1 sentences
view/table and a non-deprecated view where there are paths from that
edge to the product view that don't pass through:
a) `sentence_sessions` (i.e. this edge is not behind the new API_
b) a state-specific view for another state

Ultimately these tests are trying to answer the question: "If we feel
good about the changes in `sentence_sessions` when we flip to v2
sentences, what are all the other places that we would need to test /
that would break and aren't covered by validation on the
`sentence_sessions` dataset?". It also keeps new indirect references to
v1 sentences from being introduced going forward.

I added exemption files for each state which can be interpreted like
this:
```
US_AR_SENTENCE_V1_PRODUCT_USAGE_EXEMPTIONS: dict[
    str, dict[BigQueryAddress, dict[BigQueryAddress, set[BigQueryAddress]]]
] = {
    # This is the name of the metric export that has upstream v1 sentence dependencies
    "WORKFLOWS_FIRESTORE": {
        # The exported view with upstream v1 sentence dependencies
        US_AR_INSTITUTIONAL_WORKER_STATUS_VIEW_BUILDER.address: {
            # The deprecated v1 sentences view it depends on
            SENTENCES_PREPROCESSED_VIEW_BUILDER.address: {
                # The places where this deprecated v1 view is referenced (i.e. the call sites we need to migrate)
                US_AR_ELIGIBLE_CRIMINAL_HISTORY_309_VIEW_BUILDER.address,
                US_AR_RESIDENT_METADATA_VIEW_BUILDER.address,
            },
            ... etc
        },
        ... etc
    },
    ... etc
}
```

Some notes on the exemptions:
* There are states that only have v2 sentences ingested but still depend
on v1 `sessions` views. These legacy `sessions` views are actually
reading from v2 sentences right now so this is not explicitly a bug (I
don't think), but we WILL nee to migrate these references over to be
able to delete the legacy `sessions` views
* Any state with `INSIGHTS` imports on has a TON of bad edges due to the
fact that insights views read from the "standard" set of aggregated
metrics tables which pull in basically every metric you could imagine. I
do not think we're actually reading from any sentence-backed metrics so
we should migrate `INSIGHTS` views to read from a agg metrics collection
that includes a narrower set of metrics to remove these bad upstream
edges and make it clearer what changes actually impact `INSIGHTS`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#33402

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f3f5e9371ab1c1f5fcc88107a188827007726b0a