[US_AZ] Fixing Mandatory Literacy Tab Placement (Recidiviz/recidiviz-data#51083)
## Description of the change

We made a change a while ago that combined a criteria that checks when a
resident "meets" literacy and is "enrolled" in literacy programs, but
propagating that change into all files missed one change in our tab
placement file that decides who falls into the Almost Eligible category.
This change fixes that and makes the wordings align.

We will see 21 people showing for DTP and 200 for TPR under almost
eligibility except meeting mandatory literacy.

This has been loaded to sandbox under `shalin`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7375dc2e604613ed97e9f9a9630ef1c567659625