Hydrate new sex field for states currently hydrating state_person gender (Recidiviz/recidiviz-data#51087)
## Description of the change

Recidiviz/recidiviz-data#50987 added a new `sex` enum to state_person, allowing us to hydrate
sex and gender separately. This change hydrates sex for all states that
provide sex/gender data: in all but two cases, this was done by just
reusing the existing mappings for gender, since we don't receive
separate sex and gender data from most states. The two exceptions are
US_IX and US_NY:
- US_IX is the only state that uses the StateGender.TRANS_MALE and
StateGender.TRANS_FEMALE values. These were mapped to StateSex.FEMALE
and StateSex.MALE respectively.
- US_NY is the only state that currently provides separate columns for
sex and gender. Therefore, it's the only state with an update to its
ingest view (not just the mapping), since we can hydrate sex and gender
separately by pulling the relevant columns into the ingest view.

When this change lands, all states that currently hydrate state_person
gender will also be hydrating state_person sex. The next step will be to
update downstream views to pull from the newly-hydrated sex field
instead of gender: for all states except IX and NY, this will be a
change in name only as the fields are hydrated the same way. This will
allow us to remove the state_person gender mappings for any state in
which gender is not given explicitly, since we assume by default that
the sex/gender data we receive concerns a person's birth sex, not gender
identity.

**For more information on the migration plan, see this epic:** Recidiviz/recidiviz-data#50969

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#50971

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9e754290123dd47e26edb9d093d263790f88b402