[Insights] Remove redundant database sessions in OutliersQuerier (Recidiviz/recidiviz-data#51128)
## Description of the change

There are some code paths within OutliersQuerier where multiple methods
in the same call stack each create their own database session. This
creates several concurrent db connections, which sit around in the
connection pool until we run out of slots. This PR refactors all such
paths that I (and copilot) could find so the session is passed down
through instead.

It also wraps two potentially leaky sessions in managed contexts so they
get cleaned up. (These are only called when a user reports an incorrect
roster, so I don't think they were leaking much in practice.)

This doesn't really feel like a permanent solution: we have to manually
keep track of which methods are called by other methods that create a
session, and methods that could be called both internally or externally
are annoying. I think the real answer is to manage the Querier itself
with a context so that it can maintain one session over its lifetime. I
didn't jump directly to that for two reasons: it'll require changes
everywhere we use the querier, and it may make things more complicated
in the places where we're writing to the DB and using the session for
transaction management.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

This should help with Recidiviz/recidiviz-data#50621

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7f1fc36022253e0d6b7adc64232972fb882e9244