[US_AZ] Fix agreement form side panel notes (Recidiviz/recidiviz-data#51102)
## Description of the change

The date filter in the query fragment we are using to send agreement
form information to the FE was filtering the results to include only
those that took place at least 6 months ago, and exclude any more recent
than that. I think there were two bugs here:
1. The date filter was backwards, and was probably intended to only
include updates that happened less than 6 months ago.
2. I don't think we actually want to filter these people out of the
results. They do appear in the tool, so filtering them out of these
results will just make them appear with a NULL agreement form status. If
we included this filter to accomplish some other goal at first, then we
should still include them and point to that goal somehow in the
`note_title` for the row.

This changes the query slightly so that anytime a person did not sign
the form, regardless of how long ago that was, they will have a case
note with the `note_title: 'Not Signed'`. I kept the conditions separate
in case we want to change the copy that shows when someone did not sign
more than 6 months ago - if we want to suggest the COIII revisit the
form with a resident after that amount of time, or if this tells us
something about a person's eligibility overall, we can still indicate
that.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f124f60f6a903a1258d5ffde720bb5faeb9f66d9