[looker] Add disclaimer about state charge and state early discharge dashboard tables (Recidiviz/recidiviz-data#51204)
## Description of the change

Looker only joins a view to the root entity via a single path when
pulling data into a dashboard, which may result in missing rows for
multi-parent entities. Since our only existing multi-parent entities are
entities that are in the process of deprecation and adding support for
multi-parent entities is non-trivial, instead add a disclaimer to the
state person and normalized state person tables that state charge and
state early discharge information may be incorrect and add a check that
there are no new multiparent entities when we are building an explore.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#51203 Recidiviz/recidiviz-data#50210

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4bbf8ab93b42794ca5d02792dd5d66dd1bb9030d