[US_IX] Face-to-face contacts: tasks validation (Recidiviz/recidiviz-data#51085)
## Description of the change

This PR validates our migration of face-to-face contacts in tasks.
During validation, we observed some discrepancies, but these are
expected. The reason is that we are transitioning from a dynamic cadence
(which recalculates after each contact) to a fixed cadence based on our
existing us_xx_contact_cadence_spans infrastructure.

As noted in my comments, the discrepancies are minimal and fall within
an acceptable range. They stem from the cadence model change rather than
any errors in the migration logic. Specifically:

1. Last contact date: Matches consistently across both infrastructures,
indicating it’s being accounted for correctly.
2. Cadence applied: Matches in almost all cases. A very small minority
do not, but the mismatch rate is low enough to be acceptable.

Overall, the evidence suggests the differences are fully explained by
the dynamic vs fixed cadence change, and not by unexpected issues in the
migration.

I also created an opportunity query for IX tasks v2 and added
face-to-face to the tasks configs. It's all uploaded to a sandbox:

```
python -m recidiviz.tools.load_views_to_sandbox \
        --project_id recidiviz-staging\
        --sandbox_dataset_prefix hsalas\
        --state_code_filter US_IX manual\
        --dataset_ids_to_load tasks_views,task_eligibility_criteria_us_ix,compliance_task_eligibility_spans_us_ix,static_reference_data_views,reference_views,workflows_views,observations__person_event\
        --update_descendants False
```

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0fc27f1c697357bc7f1a70295a6fa69161059e88