[US_TN] Remove invalid supervisors exemption (Recidiviz/recidiviz-data#51461)
## Description of the change

Huge thanks to @lydiamasri2 for addressing the supervisor issues
outlined in Recidiviz/recidiviz-data#46144! Will follow with a PR that actually cleans up the
exemption mechanism now that all states are compliant (!!) and actually
hydrates staff ids on supervisor periods.

Successful pipeline run with this change here:
https://console.cloud.google.com/dataflow/jobs/us-east1/2025-10-29_09_28_56-8804665799427484217;mainTab=JOB_GRAPH;graphView=0?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#46144

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2b25abf9d45841c7cf37c8a38d199693052d8e24