[Bugfix] Use static Pathways-enabled states list in BQ views (Recidiviz/recidiviz-data#51493)
## Description of the change

In Recidiviz/recidiviz-data#51456, @justkunz surfaced that a number of pathways views appear to
have changed when loading views to a sandbox with project
`recidiviz-123`. This is happening because the views are using a helper
which determines a list of pathways-enabled states that checks the env
AT IMPORT TIME, which is before the script applies the
`local_project_id_override()`.

In order to solve this issue, I split the helper into three to reflect
what I think are the three use cases:
* `get_pathways_enabled_states_for_cloud_sql()` - Includes US_OZ ONLY if
we are running in / against staging
* `get_pathways_enabled_states_for_bigquery()` - Never include US_OZ (no
real / useful data in BQ in OZ)
* `get_pathways_enabled_states_for_local_development()` - Always include
US_OZ. This helper can be called in scripts in places where there is not
yet a project_id set.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#51456

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b462878aef17a7ad73d0aa671034ca031bad381b